✨ Prettify

+++ contracts/interfaces/IRewardEscrow.sol
@@ -18,15 +18,24 @@ interface IRewardEscrow {
     // Views
     function getKwentaAddress() external view returns (address);
 
-    function balanceOf(address account) external view returns (uint);
+    function balanceOf(address account) external view returns (uint256);
 
-    function numVestingEntries(address account) external view returns (uint);
+    function numVestingEntries(address account) external view returns (uint256);
 
-    function totalEscrowedAccountBalance(address account) external view returns (uint);
+    function totalEscrowedAccountBalance(address account)
+        external
+        view
+        returns (uint256);
 
-    function totalVestedAccountBalance(address account) external view returns (uint);
+    function totalVestedAccountBalance(address account)
+        external
+        view
+        returns (uint256);
 
-    function getVestingQuantity(address account, uint256[] calldata entryIDs) external view returns (uint, uint);
+    function getVestingQuantity(address account, uint256[] calldata entryIDs)
+        external
+        view
+        returns (uint256, uint256);
 
     function getVestingSchedules(
         address account,
@@ -40,9 +49,19 @@ interface IRewardEscrow {
         uint256 pageSize
     ) external view returns (uint256[] memory);
 
-    function getVestingEntryClaimable(address account, uint256 entryID) external view returns (uint, uint);
+    function getVestingEntryClaimable(address account, uint256 entryID)
+        external
+        view
+        returns (uint256, uint256);
 
-    function getVestingEntry(address account, uint256 entryID) external view returns (uint64, uint256, uint256);
+    function getVestingEntry(address account, uint256 entryID)
+        external
+        view
+        returns (
+            uint64,
+            uint256,
+            uint256
+        );
 
     // Mutative functions
     function vest(uint256[] calldata entryIDs) external;
